Updated Pinpoint samples based on recommendations from PR reviewers.
