+++ dotnet/example_code/PinpointEmail/pinpoint_send_email_message_email_api.cs
@@ -13,12 +13,12 @@
 */
 
 // snippet-sourcedescription:[pinpoint_send_email_message_email_api demonstrates how to send a transactional email message by using the SendEmail operation in the Amazon Pinpoint Email API.]
-// snippet-service:[Amazon Pinpoint]
+// snippet-service:[mobiletargeting]
 // snippet-keyword:[dotnet]
 // snippet-keyword:[.NET]
 // snippet-keyword:[Amazon Pinpoint Email API]
 // snippet-keyword:[Code Sample]
-// snippet-keyword:[CreateExportJob]
+// snippet-keyword:[SendEmail]
 // snippet-sourcetype:[full-example]
 // snippet-sourcedate:[2019-01-20]
 // snippet-sourceauthor:[AWS]
