+++ java/example_code/pinpoint/pinpoint_create_import_job.java
@@ -12,22 +12,29 @@
  * specific language governing permissions and limitations under the License.
 */
 
-// snippet-sourcedescription:[pinpoint_import_segment demonstrates how to create a segment by importing information about the endpoints that belong to the segment.]
+// snippet-sourcedescription:[pinpoint_create_import_job demonstrates how to import a segment of customers using a source file stored in an Amazon S3 bucket.]
 // snippet-service:[Amazon Pinpoint]
 // snippet-keyword:[Java]
 // snippet-keyword:[Amazon Pinpoint]
 // snippet-keyword:[Code Sample]
 // snippet-keyword:[CreateImportJob]
-// snippet-sourcetype:[full-example]
+// snippet-sourcetype:[snippet]
 // snippet-sourcedate:[2018-08-07]
 // snippet-sourceauthor:[AWS]
-// snippet-start:[pinpoint.java.pinpoint_import_segment.complete]
+// snippet-start:[pinpoint.java.pinpoint_create_import_job.complete]
+
+package com.amazonaws.examples.pinpoint;
 
 import com.amazonaws.AmazonServiceException;
 import com.amazonaws.regions.Regions;
 import com.amazonaws.services.pinpoint.AmazonPinpoint;
 import com.amazonaws.services.pinpoint.AmazonPinpointClientBuilder;
-import com.amazonaws.services.pinpoint.model.*;
+import com.amazonaws.services.pinpoint.model.CreateImportJobRequest;
+import com.amazonaws.services.pinpoint.model.CreateImportJobResult;
+import com.amazonaws.services.pinpoint.model.Format;
+import com.amazonaws.services.pinpoint.model.GetImportJobRequest;
+import com.amazonaws.services.pinpoint.model.GetImportJobResult;
+import com.amazonaws.services.pinpoint.model.ImportJobRequest;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import com.amazonaws.services.s3.model.AmazonS3Exception;
@@ -37,22 +44,24 @@ import java.nio.file.Paths;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
-public class ImportSegment {
+public class ImportEndpoints {
 
     public static void main(String[] args) {
 
         final String USAGE = "\n" +
-        "ImportSegment - Creates a segment by: \n" +
-        "1.) Uploading the endpoint definitions that belong to the segment to an Amazon S3 bucket. \n" +
-        "2.) Importing the endpoint definitions from the bucket to an Amazon Pinpoint application." +
-        "    Amazon Pinpoint creates a segment that has the specified name.\n\n" +
-        "Usage: ImportSegment <endpointsFileLocation> <s3BucketName> <iamImportRoleArn> <segmentName> <applicationId>\n\n" +
+        "ImportEndpoints - Adds endpoints to an Amazon Pinpoint application by: \n" +
+        "1.) Uploading the endpoint definitions to an Amazon S3 bucket. \n" +
+        "2.) Importing the endpoint definitions from the bucket to an Amazon Pinpoint " +
+                "application.\n\n" +
+        "Usage: ImportEndpoints <endpointsFileLocation> <s3BucketName> <iamImportRoleArn> " +
+                "<applicationId>\n\n" +
         "Where:\n" +
-        "  endpointsFileLocation - The relative location of the JSON file that contains the endpoint definitions.\n" +
-        "  s3BucketName - The name of the Amazon S3 bucket to upload the JSON file to. If the bucket doesn't " +
-        "exist, a new bucket is created.\n" +
-        "  iamImportRoleArn - The ARN of an IAM role that grants Amazon Pinpoint read permissions so the S3 bucket.\n" +
-        "  segmentName - The name for the segment that you are creating or updating." +
+        "  endpointsFileLocation - The relative location of the JSON file that contains the " +
+                "endpoint definitions.\n" +
+        "  s3BucketName - The name of the Amazon S3 bucket to upload the JSON file to. If the " +
+                "bucket doesn't exist, a new bucket is created.\n" +
+        "  iamImportRoleArn - The ARN of an IAM role that grants Amazon Pinpoint read " +
+                "permissions to the S3 bucket.\n" +
         "  applicationId - The ID of the Amazon Pinpoint application to add the endpoints to.";
 
         if (args.length < 1) {
@@ -63,14 +72,13 @@ public class ImportSegment {
         String endpointsFileLocation = args[0];
         String s3BucketName = args[1];
         String iamImportRoleArn = args[2];
-        String segmentName = args[3];
-        String applicationId = args[4];
+        String applicationId = args[3];
 
         Path endpointsFilePath = Paths.get(endpointsFileLocation);
         File endpointsFile = new File(endpointsFilePath.toAbsolutePath().toString());
         uploadToS3(endpointsFile, s3BucketName);
 
-        importSegment(endpointsFile.getName(), s3BucketName, iamImportRoleArn, segmentName, applicationId);
+        importToPinpoint(endpointsFile.getName(), s3BucketName, iamImportRoleArn, applicationId);
 
     }
 
@@ -102,8 +110,8 @@ public class ImportSegment {
         }
     }
 
-    private static void importSegment(String endpointsFileName, String s3BucketName, String iamImportRoleArn,
-                                      String segmentName, String applicationId) {
+    private static void importToPinpoint(String endpointsFileName, String s3BucketName,
+                                         String iamImportRoleArn, String applicationId) {
 
         // The S3 URL that Amazon Pinpoint requires to find the endpoints file.
         String s3Url = "s3://" + s3BucketName + "/imports/" + endpointsFileName;
@@ -111,11 +119,9 @@ public class ImportSegment {
         // Defines the import job that Amazon Pinpoint runs.
         ImportJobRequest importJobRequest = new ImportJobRequest()
                 .withS3Url(s3Url)
-                .withFormat(Format.JSON)
-                .withRoleArn(iamImportRoleArn)
                 .withRegisterEndpoints(true)
-                .withDefineSegment(true)
-                .withSegmentName(segmentName);
+                .withRoleArn(iamImportRoleArn)
+                .withFormat(Format.JSON);
         CreateImportJobRequest createImportJobRequest = new CreateImportJobRequest()
                 .withApplicationId(applicationId)
                 .withImportJobRequest(importJobRequest);
@@ -124,38 +130,38 @@ public class ImportSegment {
         AmazonPinpoint pinpointClient = AmazonPinpointClientBuilder.standard()
                 .withRegion(Regions.US_EAST_1).build();
 
-        System.out.format("Creating segment %s with the endpoints in %s . . .\n", segmentName, endpointsFileName);
+        System.out.format("Importing endpoints in %s to Amazon Pinpoint application %s . . .\n",
+                endpointsFileName, applicationId);
 
         try {
 
             // Runs the import job with Amazon Pinpoint.
-            CreateImportJobResult importResult = pinpointClient.createImportJob(createImportJobRequest);
-            String jobId = importResult.getImportJobResponse().getId();
+            CreateImportJobResult importResult =
+                    pinpointClient.createImportJob(createImportJobRequest);
 
-            // Checks the job status until the job completes or fails.
+            String jobId = importResult.getImportJobResponse().getId();
             GetImportJobResult getImportJobResult = null;
             String jobStatus = null;
+
+            // Checks the job status until the job completes or fails.
             do {
                 getImportJobResult = pinpointClient.getImportJob(new GetImportJobRequest()
                         .withJobId(jobId)
                         .withApplicationId(applicationId));
                 jobStatus = getImportJobResult.getImportJobResponse().getJobStatus();
                 System.out.format("Import job %s . . .\n", jobStatus.toLowerCase());
-                if (jobStatus.equals("FAILED")) {
-                    System.err.println("Failed to import segment.");
-                    System.exit(1);
-                }
-                try {
-                    TimeUnit.SECONDS.sleep(3);
-                } catch (InterruptedException e) {
-                    System.err.println(e.getMessage());
-                    System.exit(1);
-                }
-            } while (!jobStatus.equals("COMPLETED"));
+                TimeUnit.SECONDS.sleep(3);
+            } while (!jobStatus.equals("COMPLETED") && !jobStatus.equals("FAILED"));
 
-            System.out.println("Finished importing segment.");
+            if (jobStatus.equals("COMPLETED")) {
+                System.out.println("Finished importing endpoints.");
+            } else {
+                System.err.println("Failed to import endpoints.");
+                System.exit(1);
+            }
 
             // Checks for entries that failed to import.
+            // getFailures provides up to 100 of the first failed entries for the job, if any exist.
             List<String> failedEndpoints = getImportJobResult.getImportJobResponse().getFailures();
             if (failedEndpoints != null) {
                 System.out.println("Failed to import the following entries:");
@@ -164,13 +170,12 @@ public class ImportSegment {
                 }
             }
 
-        } catch (AmazonServiceException e) {
-            System.err.println(e.getErrorMessage());
+        } catch (AmazonServiceException | InterruptedException e) {
+            System.err.println(e.getMessage());
             System.exit(1);
         }
 
     }
 
 }
-
-// snippet-end:[pinpoint.java.pinpoint_import_segment.complete]
+// snippet-end:[pinpoint.java.pinpoint_create_import_job.complete]
