+++ java/example_code/pinpoint/pinpoint_update_endpoints_batch.java
